Separate module for learning agents, updated tests; update setup
