+++ setup.py
@@ -6,4 +6,5 @@ setup(name='MultiMarketABM',
       packages=['mmabm'],
       ext_modules=cythonize('**/*.pyx', compiler_directives={'language_level': 3}),
       include_dirs=[numpy.get_include()],
+      zip_safe=False,
       )
